goodwatch-graph/arangodb-poc/pyproject.toml
[project]
name = "arangodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary",
    "python-arango",
    "python-dotenv"
]
